Use a simpler, correct TRAP_IN subroutine
Summary:
I originally just grabbed an LC-3 OS image from Brian Hartman,
but @jplanck pointed out that the `IN` trap subroutine
doesn't meet the specification:
it doesn't output a prompt.
So I rewrote it to be a lot simpler,
relying on the existing `PUTS`, `GETC`, and `OUT` subroutines,
and to actually output the prompt.

Closes #5.

Test Plan:
```asm
.ORIG x3000
IN
HALT
.END
```