Add Firefox-compatible Blob API download mechanism
Summary:
This is intended to address #6, by using an API that Firefox supports.
I'm leaving in a testing bridge for now.
To use the new code, append `#blob-download` to the URL.
Unfortunately, I had to change the old code a bunch, too,
so I can't guarantee that it's bug-free.
I'm happy to roll back if this breaks something.

Test Plan:
I assembled my Pig Latin program in both Chrome and Firefox.
The downloaded object files were identical in each browser,
as were the symbol tables.
The downloaded object files also agree with
my old results from the Windows assembler.
The downloaded symbol tables agree in terms of content;
the only differences are that
the Windows assembler used tab characters instead of spaces
in some places,
and that the Windows assembler added a "Scope level 0:" header,
whatever that means.